requirements.txt
flask-cors
Flask
openai
python-dotenv


